Struct credentials::Client

pub struct Client { /* fields omitted */ }

A client which fetches secrets. Under normal circumstances, it's usually easier to use the static credentials::var and credentials::file methods instead, but you may need to use this to customize behavior.

Methods

impl Client

fn new(options: Options) -> Result<Client, Error>

Create a new client using the specified options.

fn default() -> Result<Client, Error>

Create a new client using the default options.

fn with_secretfile(secretfile: Secretfile) -> Result<Client, Error>

Create a new client using the specified Secretfile.

fn secretfile(&self) -> &Secretfile

Provide access to a copy of the Secretfile we're using.

fn var<S: AsRef<str>>(&mut self, name: S) -> Result<String, Error>

Fetch the value of an environment-variable-style credential.

fn file<S: AsRef<Path>>(&mut self, path: S) -> Result<String, Error>

Fetch the value of a file-style credential.